+++ photutils/segmentation/catalog.py
@@ -263,6 +263,9 @@ class SourceCatalog:
             raise TypeError('segment_img must be a SegmentationImage')
         if segment_img.shape != self._data.shape:
             raise ValueError('segment_img and data must have the same shape.')
+        if np.sum(segment_img.data) == 0:
+            raise ValueError('segment_img must have at least one non-zero '
+                             'label.')
         return segment_img
 
     def _validate_array(self, array, name, shape=True, dtype=None):